requirements.txt
# 核心依赖
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pillow==10.0.1
numpy==1.24.3
opencv-python==4.8.1.78

# 人脸识别和向量处理
insightface==0.7.3
onnxruntime==1.16.3
scikit-learn==1.3.0

# 向量数据库
faiss-cpu==1.7.4
chromadb==0.4.15

# 数据库和存储
sqlalchemy==2.0.23

# Web框架和API
flask==3.0.0
flask-cors==4.0.0
tmdbv3api==1.9.0

# 图像处理和下载
wget==3.2
tqdm==4.66.1

# 配置管理
python-dotenv==1.0.0
pyyaml==6.0.1

# 日志和调试
loguru==0.7.2
